tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/module/slow-cycle.html

<!DOCTYPE html>
<title>Cyclic graph with slow imports</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
import { loaded } from "./slow-module-graph-a.js";

test(() => {
  assert_true(loaded);
}, "module graph with cycles load even if part of the graph loads slow");
</script>
